+++ CHANGELOG.md
@@ -11,7 +11,8 @@
   * `ignore-files` to `exclude`
   * `maximum-nesting` to `maximum-nesting-level`
   * `number-of-arguments` to `number-of-parameters`
-* Update README
+* Update README.
+* Add static code diagnostic `member-ordering-extended`.
 
 ## 2.4.1